SFR-257 Add bulk processing for ElasticSearch indexing
With an increased volume of records coming from HathiTrust,
indexing individual records is no longer viable. This update adds
batch processing of index requests in batches of 50. This is still
on the low side for ElasticSearch and may be revised upwards at a
later time.

The considerations for batch size need to be lambda run time vs.
ElasticSearch optimization. Hypothetically batches of 500-1000 are
possible but these runtimes should be evaluated before attempting
this processing